chore: move constant block Set to Utils
